setting flutter project for frontend development
