+++ .gitignore
@@ -1,3 +1,11 @@
+# Claude Code & agents
+.claude/
+.agents/
+skills-lock.json
+
+# Project spec (local reference only)
+SPEC.md
+
 # Miscellaneous
 *.class
 *.log
