+++ assets/data/exercises.json
@@ -1,0 +1,1 @@
+[]
