+++ lib/core/theme/app_theme.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+
+class AppTheme {
+  static final dark = ThemeData.dark(useMaterial3: true);
+  static final light = ThemeData.light(useMaterial3: true);
+}
