+++ lib/core/utils/mock_data.dart
@@ -1,0 +1,1 @@
+// TODO: implement in Phase 1
